+++ rust_qsim/src/simulation/scenario/network.rs
@@ -17,7 +17,7 @@ use tracing::info;
 pub struct Network {
     nodes: IntMap<Id<Node>, Node>,
     links: IntMap<Id<Link>, Link>,
-    effective_cell_size: f32,
+    effective_cell_size: f64,
 }
 
 #[derive(Debug, Clone)]
@@ -37,9 +37,9 @@ pub struct Link {
     pub from: Id<Node>,
     pub to: Id<Node>,
     pub length: f64,
-    pub capacity: f32,
-    pub freespeed: f32,
-    pub permlanes: f32,
+    pub capacity: f64,
+    pub freespeed: f64,
+    pub permlanes: f64,
     pub modes: IntSet<Id<String>>,
     pub partition: u32,
     pub attributes: InternalAttributes,
@@ -60,7 +60,7 @@ impl Network {
         }
     }
 
-    pub fn effective_cell_size(&self) -> f32 {
+    pub fn effective_cell_size(&self) -> f64 {
         self.effective_cell_size
     }
 
@@ -227,7 +227,7 @@ impl Network {
             .collect::<Vec<&Node>>()
     }
 
-    pub fn set_effective_cell_size(&mut self, effective_cell_size: f32) {
+    pub fn set_effective_cell_size(&mut self, effective_cell_size: f64) {
         self.effective_cell_size = effective_cell_size;
     }
 
@@ -357,9 +357,9 @@ impl Link {
         from: Id<Node>,
         to: Id<Node>,
         length: f64,
-        capacity: f32,
-        freespeed: f32,
-        permlanes: f32,
+        capacity: f64,
+        freespeed: f64,
+        permlanes: f64,
         modes: IntSet<Id<String>>,
         partition: u32,
     ) -> Self {
